Closes #43 #36 (#45)
* Closes #36 - SF MongoDB bind was attempted even during udpate

DBManager initiates connection to DB by getting the DB co-orindates
from the property stored on deployment. Once db bind info is retrieved
it is then stored locally & on any future ocassions which requires
db connections to be reset again, the same bind info is used.

However during start if BOSH is down, then db bind info is not
fetched & during an update request after it completes successfully
if bind info is not found in cache its assumed that its a creation
flow and tries creating the bind once again which errors as reported.

To fix this couple of things have been done:
1. During start if unexpected errors occur then initialize of DB
manager is retried till it succeeds (with a delay)
2. Post create/update operation before creating the binding first
a check is done to see if binding already exists and only then
creation of binding is done.
3. Retry is introduced even for erros occuring during polling of
create/update task status.
4. If bind info cannot be retrieved post update/create bosh errors
even in such scenarios retry happens till it succeeds.

Hopefully with all the retries in place it must certainly fix the
DBManager failures due to BOSH errors.

Fixes #36

* increased promise delay to ensure required state during verification

* Incorporated review comments

* Fixes #43 Retrying unlock of deployment leads to a domino effect on BOSH

On unlock error, retry is implemented. However if BOSH is already down
performing unlock operation results in a CF UPDATE which internally
invokes /getDeployments on BOSH which is an expensive call. The fix is:
1. Stop retry of unlock operation in individual run of poller
2. Find deploymentName for the instance Id initially in the BOSH cache
and then if there is a cache miss goto BOSH to getDeployments
3. FabrikStatusPoller also gets all the deploymentNames from the BOSH Cache

Closes #43

* Fix jshint errors